Switch some scripts to ESM (#10867)
